// src/app/models/items.model.ts
export interface Item {
    id: number;
    deleted: boolean;
    type: string;
    by: string;
    time: number;
    text: string;
    dead: boolean;
    parent: string;
    poll: number;
    kids: number[];
    url: string;
    score: number;
    title: string;
    parts: number[];
    descendants: number;
  }